refactor(date): wire v0 date integration into framework
- Delete old DateAdapter.ts and adapters/vuetify.ts
- Move DateAdapter interface into bridge.ts (only consumer)
- Update index.ts exports: DateSymbol, VuetifyDateBridge, LegacyDateAdapterCompat, StringDateAdapter
- Update framework.ts: use date.install(app) and DateSymbol instead of manual provides
- Update date.ts import: DateAdapter from ./bridge instead of deleted ./DateAdapter
